Show toc table with multiple packages

Show toc table with multiple packages

cross.toc(pkgs, date1 = "1970-1-1", date2 = Sys.Date())

Arguments

pkgs
character vector containing the package names.

date1, date2
date range to consider (in the format "%Y-%m-%d").

Value

A data.frame with 3 columns:

Version

The package version number

Published

The date at which the specific version was published

Package

The package name

See also

Examples

cross.toc(c("magrittr", "rlang"), date1 = "2012-02-01", date2 = "2020-02-01")

#> *****************  GroundhogR's directory *********************************************************

#> <PRESS ENTER> to accept the default directory ('/Users/runner/groundhogR/') for saving packages

#> groundhogR needs a directory to use as a library for saving downloaded and installed packages.
#> <PRESS ENTER> to use the default for this computer ('/Users/runner/groundhogR/').
#> Or, type in the folder you would like to use instead (e.g, 'c:/dropbox/groundhogR'). Do not include quotes.
#> If the folder does not already exists, it will be created.
#> Type 'quit' to not choose a directory at this time

#> ************************************************************************************************

#> Please enter your reponse:

#> The folder was succesfully set to: '/Users/runner/groundhogR'

#> You can undo this selection by running: set.groundhogr.folder('')
#> 
#> 
#> 

#>       Version  Published  Package
#> 65070   1.0.0 2014-02-25 magrittr
#> 63395   1.0.1 2014-05-15 magrittr
#> 59900     1.5 2014-11-22 magrittr
#> 37243     0.1 2017-05-06    rlang
#> 36868   0.1.1 2017-05-18    rlang
#> 34345   0.1.2 2017-08-09    rlang
#> 31711   0.1.4 2017-11-05    rlang
#> 30142   0.1.6 2017-12-21    rlang
#> 28273   0.2.0 2018-02-20    rlang
#> 24508   0.2.1 2018-05-30    rlang
#> 21742   0.2.2 2018-08-16    rlang
#> 19708   0.3.0 2018-10-22    rlang
#> 19608 0.3.0.1 2018-10-25    rlang
#> 17486   0.3.1 2019-01-08    rlang
#> 14628   0.3.2 2019-03-21    rlang
#> 14346   0.3.3 2019-03-29    rlang
#> 14048   0.3.4 2019-04-07    rlang
#> 11202   0.4.0 2019-06-25    rlang
#> 7330    0.4.1 2019-10-24    rlang
#> 6328    0.4.2 2019-11-23    rlang
#> 3856    0.4.3 2020-01-24    rlang
#> 3714    0.4.4 2020-01-28    rlang