8 Final function: groundhog.library()

8 Final function: groundhog.library()

groundhog.library(
  pkg,
  date,
  quiet.install = TRUE,
  include.suggests = FALSE,
  current.deps = "Rcpp",
  ignore.package.conflicts = FALSE,
  force.source = FALSE,
  force.install = FALSE
)

Arguments

pkg
character. Name of the package to install

date
date (in the format "%Y-%m-%d"), Required date for the package.

quiet.install
Logical (defaults to TRUE).Run install.packages() with quiet = TRUE?

include.suggests
Logical (defaults to FALSE. Should suggested packages be installed?

current.deps
Dependencies that will install current version regardless of date.

ignore.package.conflicts
Logical (defaults to FALSE). With TRUE, if an already attached package will be installed, it is detached, but all depednecies are left installed. With default, FALSE, script stops and asks for session restart.

force.source
Logical (defaults to FALSE). If TRUE,` will skip CRAN and MRAN attempts, download tar.gz, and install from it.

force.install
Logical (defaults to FALSE). If TRUE, even if package is found for this R-build, it will reinstall it.

Value

a character vector containing all active packages for the session, with their version number, under the format pkg_vrs.

Examples

groundhog.library("magrittr", "2018-02-12")

#> groundhog.library() says [using R-4.0.0]:

#> ---------------- R Version check ----------------

#> You are using  R : '4.0.0, but for the date you entered (2018-02-12),
#> the current R was: '3.4.3'

#> This version mismatch has two consequences:
#> 1) Some of the packages you install, and their dependencies, probably will take longer to install, possibly *minutes* longer.
#> 2) There is some chance the original code won't run in the newer R version

#> 
#> Bottom line

#> It's probably worth continuing as-is, and if you get tired of waiting for packages to install,
#> or get errors trying to run it, then you try it with the appropriate version of R. 
#> Running older R is actually quite easy. To get instructions, execute:
#> msg.R.switch('2018-02-12') 

#> 
#> 
#> Regardless of your choice, this warning will not appear again within the next hour

#> Type 'Q', 'quit' or 'stop' to stop the script.
#> Anything else to continue

#> 

#> You typed '' the script continues...

#> groundhog.library() says [using R-4.0.0]:

#> Loading magrittr_1.5 requires loading 1 packages, of which 1 will need to be installed.

#> 
#> 
#> 
#> groundhog.library() says: Installing 'magrittr_1.5', package #1 out of 1 needed for 'magrittr_1.5'

#> > The time now is 14:24, and we quite roughly estimate the process to end around 14:24
#> > It is unlikely to finish after 14:24
#> > These estimates will be revised after each package installs, but they will remain noisy estimates.
#> > Installation is slow because you are using R-4.0.0
#> > If you run this script with the R version available on the date you entered: '2018-02-12', i.e., R-3.4.3,
#>    the entire installation would take about a minute or two.
#> > Instructions for running older version of R:  https://grondhogR.com/olderR

#> Installing Binary from CRAN

#> magrittr_1.5 installed succesfully. Saved to: /Users/runner/groundhogR/R-4.0.0/magrittr_1.5

#> >Successfully loaded magrittr_1.5